Fix cascading compilation errors from model/interface changes, update SharedInterop, MetadataHub, and MetadataBroadcaster
Agent-Logs-Url: https://github.com/GooGuTeam/osu-server-spectator/sessions/fed5a5cd-8fc3-49ea-ac42-ef2d727bb19b

+++ osu.Server.Spectator/Database/DatabaseAccess.cs
@@ -556,7 +556,7 @@ namespace osu.Server.Spectator.Database
         {
             return Task.FromResult<osu_build?>(new osu_build
             {
-                build_id = buildId,
+                build_id = (uint)buildId,
                 allow_bancho = true,
                 hash = null,
                 users = 0,
